Moving StorageBackend and related artifacts to their proper place
This PR is exclusively about moving files around, no code change.

changelog_begin
changelog_end
