+++ ledger/participant-integration-api/src/main/scala/platform/indexer/JdbcIndexer.scala
@@ -19,8 +19,9 @@ import com.daml.platform.ApiOffset.ApiOffsetConverter
 import com.daml.platform.common
 import com.daml.platform.common.MismatchException
 import com.daml.platform.configuration.ServerRole
-import com.daml.platform.indexer.parallel.{ParallelIndexerFactory, PostgresStorageBackend}
+import com.daml.platform.indexer.parallel.ParallelIndexerFactory
 import com.daml.platform.store.appendonlydao.events.{CompressionStrategy, LfValueTranslation}
+import com.daml.platform.store.backend.postgresql.PostgresStorageBackend
 import com.daml.platform.store.dao.{JdbcLedgerDao, LedgerDao}
 import com.daml.platform.store.{DbType, FlywayMigrations, LfValueTranslationCache}
